+++ R/mitemap_tools.R
@@ -1,14 +1,35 @@
 #' Rename files with number inside a folder
+#' 
+#' @description
+#' MiteMap software may export zip files with parenthesis and space in the name
+#'  (for example "My file (1).zip", which can be problematic.
+#'  This function rename files by adding the number inside the parenthesis 
+#'  before the extension (for example "My_file_1.zip"). It also rename the files
+#'  inside the zip file to avoid confusion. Note that the name of the csv inside
+#'  a zip file with a parenthesis do not have parenthesis into this name. This 
+#'  function add the number at the end of the csv file in order to match to
+#'  the zip file name.
+#'  
+#'  If `keep_original` is TRUE, the original zip files with parenthesis are kept.
+#'   
+#'  Note that this renaming must be also done in the metadata files when using
+#'  the [import_mitemap()] function. 
 #'
 #' @param path_to_folder (required) Path to folder
-#' @param remove_zip_files_with_parenthesis (logical, default TRUE). Do we
-#'   remove zip_files with parenthesis in the name?
+#' @param keep_original (logical, default FALSE). Do we keep zip_files with
+#'  parenthesis in the name?
 #'
-#' @returns Nothing. Only modify files in a folder
+#' @returns Nothing. Only modify files in a folder.
 #' @author Adrien Taudière
 #' @export
 #'
-rename_files_with_number <- function(path_to_folder, remove_zip_files_with_parenthesis = TRUE) {
+#' @examples
+#' dontrun{
+#' rename_files_with_number(system.file("extdata", "mitemap_example", package = "MiteMapTools"))
+#' rename_files_with_number(system.file("extdata", "mitemap_example_large", package = "MiteMapTools"),
+#' keep_original=TRUE)
+#' }
+rename_files_with_number <- function(path_to_folder, keep_original = FALSE) {
   old_wd <- getwd()
   folders_to_renamed <- list.files(here::here(path_to_folder),
     pattern = "\\([[:xdigit:]]\\)",
@@ -34,7 +55,7 @@ rename_files_with_number <- function(path_to_folder, remove_zip_files_with_paren
     zip::zip(sub("\\)", "", sub(" \\(", "_", folder)), files = list.files())
     setwd(old_wd)
     unlink("unzip_temp", recursive = TRUE, force = TRUE)
-    if (remove_zip_files_with_parenthesis) {
+    if (!keep_original) {
       lapply(folders_to_renamed, unlink)
     }
   }
@@ -67,4 +88,3 @@ get_file_extension <- function(file_path) {
   file_ext <- strsplit(basename(file_path), ".", fixed = TRUE)[[1]][-1]
   return(file_ext)
 }
-