+++ R/vioplot_mitemap.R
@@ -3,7 +3,9 @@
 #' @param MiteMap (required) The result of import_mitemap ($resulting_data) for raw_data
 #' @param factor : A name of column present in the MiteMap to separate violin plot.
 #' @param wrap : A name of column present in the MiteMap to wrap violin plot.
-#'
+#' @param prop_points (numeric between 0 and 1, default=NULL) If not NULL,
+#'  a proportion of points is randomly sampled and added to the violin plot.
+#'  This can help to visualize the data distribution, especially with large datasets.
 #' @return A ggplot object
 #' @export
 #' @author Adrien Taudière