+++ tests/testthat/test-vioplot_mitemap.R
@@ -1,5 +1,3 @@
-MM_filtered_centered <- filter_mitemap(MM_data, center_x = -20, center_y = -20)
-
 test_that("vioplot_mitemap works", {
-  expect_s3_class(vioplot_mitemap(MM_filtered_centered), "ggplot")
+  expect_s3_class(vioplot_mitemap(MM_data, "Treatment"), "ggplot")
 })