+++ tests/testthat/test_import_mitemap.R
@@ -1,9 +1,14 @@
-mm_csv <- suppressWarnings(import_mitemap(
-  system.file("extdata", "POUL6", package = "MiteMapTools")
-))
-dim(mm_csv$resulting_data)
+mm_csv <- import_mitemap(
+  system.file("extdata", "mitemap_example", package = "MiteMapTools"),
+  file_name_column = "File (mite ID)", verbose = FALSE
+)
+
+mm_csv_log <- import_mitemap(
+  system.file("extdata", "mitemap_example", package = "MiteMapTools"),
+  file_name_column = "File (mite ID)", verbose = FALSE, return_with_logs = TRUE
+)
 
 test_that("import_mitemap is running well", {
-  expect_equal(length(mm_csv), 4)
-  expect_equal(dim(mm_csv$resulting_data), c(117721, 12))
+  expect_equal(dim(mm_csv), c(70828, 35))
+  expect_equal(length(mm_csv_log), 4)
 })